ai-portrait-maker/requirements.txt
google-generativeai==0.3.2
streamlit==1.29.0
pillow==10.1.0
python-dotenv==1.0.0

